Re-organise date-time dir to have component subdirs
