+++ packages/components/src/date-time/date/index.tsx
@@ -20,7 +20,7 @@ import { isRTL, _n, sprintf } from '@wordpress/i18n';
  * Internal dependencies
  */
 import { getMomentDate } from './utils';
-import type { DatePickerDayProps, DatePickerProps } from './types';
+import type { DatePickerDayProps, DatePickerProps } from '../types';
 
 const TIMEZONELESS_FORMAT = 'YYYY-MM-DDTHH:mm:ss';
 const ARIAL_LABEL_TIME_FORMAT = 'dddd, LL';
