+++ packages/components/src/date-time/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Internal dependencies
+ */
+import { default as DatePicker } from './date';
+import { default as TimePicker } from './time';
+import { default as DateTimePicker } from './date-time';
+
+export { DatePicker, TimePicker };
+export default DateTimePicker;
